requirements.txt
asgiref==3.3.4
coverage==5.5
Django==3.2.2
numpy==1.20.2
Pillow==8.2.0
pytz==2021.1
smartcrop==0.3.2
sqlparse==0.4.1
